% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_java.R
\name{check_java_installation}
\alias{check_java_installation}
\title{Check if a Java installation is valid}
\usage{
check_java_installation(java_dir)
}
\arguments{
\item{java_dir}{}
}
\value{
TRUE if the Java installation is valid, otherwise stops with an error.
}
\description{
Check if a Java installation is valid
}
\examples{
check_java_installation("/path/to/java")
}
