% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_java.R
\name{check_java_version_at_path}
\alias{check_java_version_at_path}
\title{Check installed Java ver at path}
\usage{
check_java_version_at_path(java_path)
}
\arguments{
\item{java_dir}{}
}
\description{
Check installed Java ver at path
}
