% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_env.R
\name{check_java_version_cmd}
\alias{check_java_version_cmd}
\title{Check installed Java version using terminal commands}
\usage{
check_java_version_cmd(java_home = NULL)
}
\arguments{
\item{java_home}{Path to Java home directory. If NULL, the function uses the JAVA_HOME environment variable.}
}
\value{
Java version, otherwise stops with an error.
}
\description{
Check installed Java version using terminal commands
}
\examples{
\dontrun{
check_java_version_cmd()
}
}
