% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utilities.R
\name{._find_extracted_dir}
\alias{._find_extracted_dir}
\title{Find the actual extracted directory, ignoring hidden/metadata files}
\usage{
._find_extracted_dir(temp_dir)
}
\arguments{
\item{temp_dir}{The directory where files were extracted.}
}
\value{
The path to the first non-hidden directory found.
}
\description{
Find the actual extracted directory, ignoring hidden/metadata files
}
\keyword{internal}
