% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_env.R
\name{.check_java_system2}
\alias{.check_java_system2}
\title{Check and print Java path and version}
\usage{
.check_java_system2()
}
\value{
TRUE if successful, otherwise stops with an error.
}
\description{
This function checks the Java executable path and retrieves the Java version,
then prints these details to the console.
}
