% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate-deactivate.R
\name{.read_rjavaenv_config}
\alias{.read_rjavaenv_config}
\title{Read the rJavaEnv configuration file}
\usage{
.read_rjavaenv_config(project_dir = ".")
}
\arguments{
\item{project_dir}{The project directory.}
}
\description{
Read the rJavaEnv configuration file
}
\keyword{internal}
