% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate-deactivate.R
\name{.rje_clean_project}
\alias{.rje_clean_project}
\title{Clean up the project directory}
\usage{
.rje_clean_project(project)
}
\arguments{
\item{project}{The project directory.}
}
\description{
Clean up the project directory
}
\keyword{internal}
