% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate-deactivate.R
\name{.rje_comment_out_in_rprofile}
\alias{.rje_comment_out_in_rprofile}
\title{Comment out the rJavaEnv code in .Rprofile}
\usage{
.rje_comment_out_in_rprofile(project)
}
\arguments{
\item{project}{The project directory.}
}
\description{
Comment out the rJavaEnv code in .Rprofile
}
\keyword{internal}
