% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate-deactivate.R
\name{.rje_modify_rprofile}
\alias{.rje_modify_rprofile}
\title{Modify the .Rprofile to source the activate.R script}
\usage{
.rje_modify_rprofile(project)
}
\arguments{
\item{project}{The project directory.}
}
\description{
Modify the .Rprofile to source the activate.R script
}
\keyword{internal}
