% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate-deactivate.R
\name{.rje_restart_session}
\alias{.rje_restart_session}
\title{Restart the R session}
\usage{
.rje_restart_session(project, reason)
}
\arguments{
\item{project}{The project directory.}

\item{reason}{The reason for restarting the session.}
}
\description{
Restart the R session
}
\keyword{internal}
