% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate-deactivate.R
\name{.rje_write_activate_script}
\alias{.rje_write_activate_script}
\title{Write the activate.R script}
\usage{
.rje_write_activate_script(project)
}
\arguments{
\item{project}{The project directory.}
}
\description{
Write the activate.R script
}
\keyword{internal}
