% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate-deactivate.R
\name{.rje_write_infrastructure}
\alias{.rje_write_infrastructure}
\title{Write the necessary infrastructure to activate rJavaEnv}
\usage{
.rje_write_infrastructure(project)
}
\arguments{
\item{project}{The project directory.}
}
\description{
Write the necessary infrastructure to activate rJavaEnv
}
\keyword{internal}
