% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_env.R
\name{.set_and_check_java_home}
\alias{.set_and_check_java_home}
\title{Helper function to set JAVA_HOME and check if java executable exists}
\usage{
.set_and_check_java_home(java_home = NULL)
}
\arguments{
\item{java_home}{The path to the desired JAVA_HOME. If NULL, uses the current JAVA_HOME environment variable.}
}
\description{
Helper function to set JAVA_HOME and check if java executable exists
}
\keyword{internal}
