% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate-deactivate.R
\name{.set_java_env_from_config}
\alias{.set_java_env_from_config}
\title{Set the JAVA_HOME environment variable from the configuration file}
\usage{
.set_java_env_from_config(project_dir = ".")
}
\arguments{
\item{project_dir}{The project directory.}
}
\description{
Set the JAVA_HOME environment variable from the configuration file
}
\keyword{internal}
