% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_env.R
\name{.set_java_home_in_rprofile}
\alias{.set_java_home_in_rprofile}
\title{Update the .Rprofile file in the project directory}
\usage{
.set_java_home_in_rprofile(java_home)
}
\arguments{
\item{java_home}{The path to the desired JAVA_HOME.}
}
\description{
Update the .Rprofile file in the project directory
}
\keyword{internal}
