% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_env.R
\name{.set_java_paths}
\alias{.set_java_paths}
\title{Set the JAVA_HOME and PATH environment variables}
\usage{
.set_java_paths(java_home)
}
\arguments{
\item{java_home}{The path to the desired JAVA_HOME.}
}
\description{
Set the JAVA_HOME and PATH environment variables
}
\keyword{internal}
