% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate-deactivate.R
\name{.write_rjavaenv_config}
\alias{.write_rjavaenv_config}
\title{Write the rJavaEnv configuration file}
\usage{
.write_rjavaenv_config(java_home, project_dir = ".")
}
\arguments{
\item{java_home}{The Java home path.}

\item{project_dir}{The project directory.}
}
\description{
Write the rJavaEnv configuration file
}
\keyword{internal}
