% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_java.R
\name{download_java}
\alias{download_java}
\title{Download a Java distribution}
\usage{
download_java(
  version = 21,
  distribution = "Corretto",
  dest_dir = tools::R_user_dir("rJavaEnv", which = "cache"),
  platform = .detect_platform()$os,
  arch = .detect_platform()$arch,
  verbose = TRUE
)
}
\arguments{
\item{version}{The Java version to download. If not specified, defaults to the latest LTS version.}

\item{distribution}{The Java distribution to download. If not specified, defaults to "Corretto".}

\item{dest_dir}{The destination directory to download the Java distribution to. Defaults to a user-specific data directory.}

\item{platform}{The platform for which to download the Java distribution. Defaults to the current platform.}

\item{arch}{The architecture for which to download the Java distribution. Defaults to the current architecture.}

\item{verbose}{Whether to print out information about the detected platform and architecture. Defaults to TRUE.}
}
\value{
The path to the downloaded Java distribution file.
}
\description{
Download a Java distribution
}
\examples{
\dontrun{
download_java(version = "17", distribution = "Corretto")
download_java(distribution = "Corretto")
download_java()
}
}
