% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_download.R
\name{get_java_archive_extension}
\alias{get_java_archive_extension}
\title{Get extension for platform}
\usage{
get_java_archive_extension(platform)
}
\description{
Get extension for platform
}
\keyword{internal}
