% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_params.R
\name{global_quiet_param}
\alias{global_quiet_param}
\title{Global Quiet Parameter}
\usage{
global_quiet_param(quiet)
}
\arguments{
\item{quiet}{A \code{logical} value indicating whether to suppress messages. Can be \code{TRUE} or \code{FALSE}.}
}
\description{
Documentation for the \code{quiet} parameter, used globally.
}
\keyword{internal}
