% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_params.R
\name{global_sdkman_references}
\alias{global_sdkman_references}
\title{SDKMAN References}
\usage{
global_sdkman_references()
}
\description{
Standard references for SDKMAN attribution.
}
\references{
SDKMAN! - The Software Development Kit Manager: \url{https://github.com/sdkman}
}
\keyword{internal}
