% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_java.R
\name{install_java}
\alias{install_java}
\title{Install Java from a distribution file}
\usage{
install_java(java_path, project = NULL, autoset_java_path = TRUE)
}
\arguments{
\item{java_path}{The path to the Java distribution file.}

\item{project}{The project directory where Java should be installed. Defaults to the current working directory.}

\item{autoset_java_path}{Whether to set the JAVA_HOME and PATH environment variables to the installed Java directory. Defaults to TRUE.}
}
\value{
The path to the installed Java directory.
}
\description{
Install Java from a distribution file
}
\examples{
\dontrun{
install_java("path/to/any-java-17-aarch64-macos-jdk.tar.gz")
}
}
