% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_build_env.R
\name{java_build_env_set}
\alias{java_build_env_set}
\title{Set up the environment for building R packages with Java dependencies from source}
\usage{
java_build_env_set(
  java_home = Sys.getenv("JAVA_HOME"),
  where = c("session", "project", "both"),
  project_path = NULL,
  quiet = FALSE
)
}
\arguments{
\item{java_home}{The path to the desired \code{JAVA_HOME}. Defaults to the value of the \code{JAVA_HOME} environment variable.}

\item{where}{Where to set the build environment: "session", "project", or "both". Defaults to "session". When "both" or "project" is selected, the function updates the .Rprofile file in the project directory.}

\item{project_path}{The path to the project directory, required when \code{where} is "project" or "both". Defaults to the current working directory.}

\item{quiet}{A \code{logical} value indicating whether to suppress messages. Can be \code{TRUE} or \code{FALSE}.}
}
\value{
Invisibly returns \code{NULL} after setting the environment variables.
}
\description{
This function configures the current R session with the necessary environment
variables to compile Java-dependent packages like 'rJava' from source.
}
\examples{
\dontrun{
# Download and install Java 17
java_17_distrib <- java_download(version = "17", temp_dir = TRUE)
java_home_path <- java_install(
  java_distrib_path = java_17_distrib,
  project_path = tempdir(),
  autoset_java_env = FALSE # Manually set env
)

# Set up the build environment in the current session
java_build_env_set(java_home = java_home_path)

# Now, install rJava from source
install.packages("rJava", type = "source", repos = "https://cloud.r-project.org")
}
}
