% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_build_env.R
\name{java_build_env_set_rprofile}
\alias{java_build_env_set_rprofile}
\title{Helper function to write build environment settings to .Rprofile}
\usage{
java_build_env_set_rprofile(java_home, project_path)
}
\arguments{
\item{java_home}{The path to the desired \code{JAVA_HOME}.}

\item{project_path}{The path to the project directory.}
}
\description{
Helper function to write build environment settings to .Rprofile
}
\keyword{internal}
