% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_distributions_cache.R
\name{java_cache_list}
\alias{java_cache_list}
\title{List the contents of the Java distributions cache folder}
\usage{
java_cache_list(cache_dir = tools::R_user_dir("rJavaEnv", which = "cache"))
}
\arguments{
\item{cache_dir}{The cache directory to list. Defaults to the user-specific data directory.}
}
\value{
A character vector with the contents of the cache directory.
}
\description{
List the contents of the Java distributions cache folder
}
\examples{
java_cache_list()
}
