% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utilities.R
\name{java_check_current_rjava_version}
\alias{java_check_current_rjava_version}
\title{Check the Java version currently used by the loaded rJava package}
\usage{
java_check_current_rjava_version()
}
\value{
A character string representing the major Java version (e.g., "17", "21", "8"),
or \code{NULL} if \code{rJava} is not loaded or the version/property could not be retrieved.
}
\description{
Safely queries the \code{rJava} package (if loaded) to determine which Java version
it is actually using.
}
\keyword{internal}
