% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_manage_distrib_cache.R
\name{java_clear_distrib}
\alias{java_clear_distrib}
\title{Clear the Java distributions cache folder}
\usage{
java_clear_distrib(
  cache_path = getOption("rJavaEnv.cache_path"),
  check = TRUE,
  delete_all = FALSE
)
}
\arguments{
\item{cache_path}{The destination directory to download the Java distribution to. Defaults to a user-specific data directory.}

\item{check}{Whether to list the contents of the cache directory before clearing it. Defaults to TRUE.}

\item{delete_all}{Whether to delete all items without prompting. Defaults to FALSE.}
}
\value{
A message indicating whether the cache was cleared or not.
}
\description{
Clear the Java distributions cache folder
}
