% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_manage_project.R
\name{java_clear_in_project}
\alias{java_clear_in_project}
\title{Clear the Java versions symlinked in the current project}
\usage{
java_clear_in_project(project_path = NULL, check = TRUE, delete_all = FALSE)
}
\arguments{
\item{project_path}{The project directory to clear. Defaults to the current working directory.}

\item{check}{Whether to list the symlinked Java versions before clearing them. Defaults to TRUE.}

\item{delete_all}{Whether to delete all symlinks without prompting. Defaults to FALSE.}
}
\value{
A message indicating whether the symlinks were cleared or not.
}
\description{
Clear the Java versions symlinked in the current project
}
\keyword{internal}
