% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_manage_installed_cache.R
\name{java_clear_installed}
\alias{java_clear_installed}
\title{Clear the Java installations cache folder}
\usage{
java_clear_installed(
  check = TRUE,
  delete_all = FALSE,
  cache_path = getOption("rJavaEnv.cache_path")
)
}
\arguments{
\item{check}{Whether to list the contents of the cache directory before clearing it. Defaults to TRUE.}

\item{delete_all}{Whether to delete all items without prompting. Defaults to FALSE.}

\item{cache_path}{The destination directory to download the Java distribution to. Defaults to a user-specific data directory.}
}
\value{
A message indicating whether the cache was cleared or not.
}
\description{
Clear the Java installations cache folder
}
