% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_manage_installed_cache.R
\name{java_clear_installed_cache}
\alias{java_clear_installed_cache}
\title{Clear the Java installations cache folder}
\usage{
java_clear_installed_cache(
  check = TRUE,
  delete_all = FALSE,
  cache_path = getOption("rJavaEnv.cache_path")
)
}
\arguments{
\item{check}{Whether to list the contents of the cache directory before clearing it. Defaults to TRUE.}

\item{delete_all}{Whether to delete all installations without prompting. Defaults to FALSE.}

\item{cache_path}{The cache directory to clear. Defaults to the user-specific data directory.}
}
\value{
A message indicating whether the cache was cleared or not.
}
\description{
Clear the Java installations cache folder
}
\examples{
\dontrun{
java_clear_installed_cache()
}
}
