% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_manage_project.R
\name{java_clear_project}
\alias{java_clear_project}
\title{Clear the Java versions symlinked in the current project}
\usage{
java_clear_project(project_path = NULL, check = TRUE, delete_all = FALSE)
}
\arguments{
\item{project_path}{The project directory to clear. Defaults to the current working directory.}

\item{check}{Whether to list the contents of the cache directory before clearing it. Defaults to TRUE.}

\item{delete_all}{Whether to delete all items without prompting. Defaults to FALSE.}
}
\value{
A message indicating whether the symlinks were cleared or not.
}
\description{
Clear the Java versions symlinked in the current project
}
\examples{
if (interactive()) {
  java_clear_project()
}

}
