% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_env.R
\name{java_env_set_rprofile}
\alias{java_env_set_rprofile}
\title{Update the .Rprofile file in the project directory}
\usage{
java_env_set_rprofile(java_home, project_path = NULL, setup_build_env = FALSE)
}
\arguments{
\item{java_home}{The path to the desired JAVA_HOME.}

\item{project_path}{Path to the project directory.}

\item{setup_build_env}{Logical. If TRUE, also sets build-specific variables.}
}
\description{
Update the .Rprofile file in the project directory
}
\keyword{internal}
