% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_env.R
\name{java_env_unset}
\alias{java_env_unset}
\title{Unset the JAVA_HOME and PATH environment variables in the project .Rprofile}
\usage{
java_env_unset(verbose = TRUE)
}
\arguments{
\item{verbose}{Whether to print detailed messages. Defaults to TRUE.}
}
\value{
Nothing. Removes the JAVA_HOME and PATH environment variables settings from the project .Rprofile.
}
\description{
Unset the JAVA_HOME and PATH environment variables in the project .Rprofile
}
\examples{
\dontrun{
java_env_unset()
}
}
