% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_env.R
\name{java_env_unset_rprofile}
\alias{java_env_unset_rprofile}
\title{Helper function to unset JAVA_HOME in the .Rprofile}
\usage{
java_env_unset_rprofile(where, verbose = TRUE)
}
\arguments{
\item{where}{Where to remove the JAVA_HOME: "project" or "global".}

\item{verbose}{Whether to print detailed messages. Defaults to TRUE.}
}
\description{
Helper function to unset JAVA_HOME in the .Rprofile
}
\keyword{internal}
