% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_env.R
\name{java_env_unset_session}
\alias{java_env_unset_session}
\title{Helper function to unset JAVA_HOME and PATH for the current session}
\usage{
java_env_unset_session(verbose = TRUE)
}
\arguments{
\item{verbose}{Whether to print detailed messages. Defaults to TRUE.}
}
\description{
Helper function to unset JAVA_HOME and PATH for the current session
}
\keyword{internal}
