% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_env.R
\name{java_get_home}
\alias{java_get_home}
\title{Get JAVA_HOME}
\usage{
java_get_home()
}
\value{
The value of the JAVA_HOME environment variable.
}
\description{
Get the current JAVA_HOME environment variable.
}
\examples{
java_get_home()
}
