% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_manage_distrib_cache.R
\name{java_list_distrib}
\alias{java_list_distrib}
\title{List the contents of the Java distributions cache folder}
\usage{
java_list_distrib(
  cache_path = getOption("rJavaEnv.cache_path"),
  output = c("data.frame", "vector"),
  quiet = TRUE
)
}
\arguments{
\item{cache_path}{The destination directory to download the Java distribution to. Defaults to a user-specific data directory.}

\item{output}{The format of the output: \verb{data.frame`` or }vector``. Defaults to \code{data.frame}.}

\item{quiet}{A \code{logical} value indicating whether to suppress messages. Can be \code{TRUE} or \code{FALSE}.}
}
\value{
A data frame or character vector with the contents of the cache directory.
}
\description{
List the contents of the Java distributions cache folder
}
