% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_manage_distrib_cache.R
\name{java_list_distrib_cache}
\alias{java_list_distrib_cache}
\title{List the contents of the Java distributions cache folder}
\usage{
java_list_distrib_cache(
  output = c("data.frame", "vector"),
  verbose = FALSE,
  cache_dir = tools::R_user_dir("rJavaEnv", which = "cache")
)
}
\arguments{
\item{output}{The format of the output: "data.frame" or "vector". Defaults to "data.frame".}

\item{verbose}{Whether to print detailed messages. Defaults to FALSE.}

\item{cache_dir}{The cache directory to list. Defaults to the user-specific data directory.}
}
\value{
A character vector with the contents of the cache directory.
}
\description{
List the contents of the Java distributions cache folder
}
\examples{
java_list_distrib_cache()

}
