% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_manage_project.R
\name{java_list_in_project}
\alias{java_list_in_project}
\title{List the Java versions symlinked in the current project}
\usage{
java_list_in_project(
  project_path = NULL,
  output = c("data.frame", "vector"),
  quiet = TRUE
)
}
\arguments{
\item{project_path}{The project directory to list. Defaults to the current working directory.}

\item{output}{The format of the output: "data.frame" or "vector". Defaults to "data.frame".}

\item{quiet}{A \code{logical} value indicating whether to suppress messages. Can be \code{TRUE} or \code{FALSE}.}
}
\value{
A data frame or character vector with the symlinked Java versions in the project directory.
}
\description{
List the Java versions symlinked in the current project
}
