% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_env.R
\name{java_set_env}
\alias{java_set_env}
\title{Set the JAVA_HOME and PATH environment variables to a given path}
\usage{
java_set_env(java_home, where = c("both", "rsession", "project"))
}
\arguments{
\item{java_home}{The path to the desired JAVA_HOME.}

\item{where}{Where to set the JAVA_HOME: "session", "project", or "both". Defaults to "both". When "both" or "project" is selected, the function updates the .Rprofile file in the project directory to set the JAVA_HOME and PATH environment variables at the start of the R session.}
}
\value{
Nothing. Sets the JAVA_HOME and PATH environment variables.
}
\description{
Set the JAVA_HOME and PATH environment variables to a given path
}
