% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utilities.R
\name{java_urls_load}
\alias{java_urls_load}
\title{Load Java URLs from JSON file}
\usage{
java_urls_load()
}
\value{
A list with the Java URLs structured as in the JSON file by distribution, platform, and architecture.
}
\description{
Load Java URLs from JSON file
}
\keyword{internal}
