% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_valid_versions.R
\name{java_valid_major_versions_temurin}
\alias{java_valid_major_versions_temurin}
\title{Get Available Online Versions of Adoptium Temurin}
\usage{
java_valid_major_versions_temurin(arch = NULL, platform = NULL)
}
\description{
Get Available Online Versions of Adoptium Temurin
}
\keyword{internal}
