% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_valid_versions.R
\name{java_valid_major_versions_zulu}
\alias{java_valid_major_versions_zulu}
\title{Get Available Online Versions of Azul Zulu}
\usage{
java_valid_major_versions_zulu(arch = NULL, platform = NULL, imageType = "jdk")
}
\description{
Get Available Online Versions of Azul Zulu
}
\keyword{internal}
