% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_env.R
\name{java_version_check_rjava}
\alias{java_version_check_rjava}
\title{Check Java Version with a Specified JAVA_HOME Using a Separate R Session}
\usage{
java_version_check_rjava(java_home = NULL, verbose = TRUE)
}
\arguments{
\item{java_home}{The path to the desired JAVA_HOME. If NULL, uses the current JAVA_HOME environment variable.}

\item{verbose}{Logical indicating whether to print detailed messages. Defaults to TRUE.}
}
\description{
This function sets the JAVA_HOME environment variable, initializes the JVM using rJava,
and prints the Java version that would be used if the user sets the given JAVA_HOME
in the current R session. This check is performed in a separate R session to avoid
having to reload the current R session. The reason for this is that once Java is initialized in an R session,
it cannot be uninitialized unless the current R session is restarted.
}
\examples{
\dontrun{
java_version_check_rjava("/new/path/to/java")
}
}
