% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utilities.R
\name{java_version_check_rscript}
\alias{java_version_check_rscript}
\title{Check Java version using rJava}
\usage{
java_version_check_rscript(java_home)
}
\arguments{
\item{java_home}{}
}
\value{
A message with the Java version or an error message.
}
\description{
Check Java version using rJava
}
\keyword{internal}
