% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_distributions_cache.R
\name{list_java_distributions_cache}
\alias{list_java_distributions_cache}
\title{List the contents of the Java distributions cache folder}
\usage{
list_java_distributions_cache(
  cache_dir = tools::R_user_dir("rJavaEnv", which = "cache")
)
}
\arguments{
\item{cache_dir}{The cache directory to list. Defaults to the user-specific data directory.}
}
\value{
A character vector with the contents of the cache directory.
}
\description{
List the contents of the Java distributions cache folder
}
\examples{
list_java_distributions_cache()
}
