% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utilities.R
\name{platform_detect}
\alias{platform_detect}
\title{Detect platform and architecture}
\usage{
platform_detect(verbose = FALSE)
}
\arguments{
\item{verbose}{Whether to print detailed messages. Defaults to FALSE.}
}
\value{
A list of length 2 with the detected platform and architecture.
}
\description{
Detect platform and architecture
}
\keyword{internal}
