% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_corretto.R
\name{resolve_corretto_metadata}
\alias{resolve_corretto_metadata}
\title{Resolve Corretto metadata from AWS JSON}
\usage{
resolve_corretto_metadata(version, platform, arch)
}
\arguments{
\item{version}{Major Java version}

\item{platform}{Platform OS}

\item{arch}{Architecture}
}
\value{
A java_build object
}
\description{
Resolves download metadata for Amazon Corretto from the official JSON index
}
\keyword{internal}
