% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_temurin.R
\name{resolve_temurin_metadata}
\alias{resolve_temurin_metadata}
\title{Resolve Temurin metadata from Adoptium API}
\usage{
resolve_temurin_metadata(version, platform, arch)
}
\arguments{
\item{version}{Major Java version}

\item{platform}{Platform OS}

\item{arch}{Architecture}
}
\value{
A java_build object
}
\description{
Resolves download metadata for Eclipse Temurin from the Adoptium API
}
\keyword{internal}
