% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_zulu.R
\name{resolve_zulu_metadata}
\alias{resolve_zulu_metadata}
\title{Resolve Zulu metadata from Azul API}
\usage{
resolve_zulu_metadata(version, platform, arch)
}
\arguments{
\item{version}{Major Java version}

\item{platform}{Platform OS}

\item{arch}{Architecture}
}
\value{
A java_build object
}
\description{
Resolves download metadata for Azul Zulu from the Azul API
}
\keyword{internal}
