% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_params.R
\name{rjava_path_locking_note}
\alias{rjava_path_locking_note}
\title{rJava Path-Locking Documentation}
\usage{
rjava_path_locking_note()
}
\description{
Documentation template for rJava path-locking behavior.
}
\section{rJava Path-Locking}{

\strong{Important for rJava Users}: This function sets environment variables
(JAVA_HOME, PATH) that affect both command-line Java tools and rJava initialization.
However, due to rJava's path-locking behavior when \code{\link[rJava]{.jinit}} is called
(see \url{https://github.com/s-u/rJava/issues/25}, \url{https://github.com/s-u/rJava/issues/249}, and \url{https://github.com/s-u/rJava/issues/334}),
this function must be called \strong{BEFORE} \code{\link[rJava]{.jinit}} is invoked. Once \code{\link[rJava]{.jinit}}
initializes, the Java version is locked for that R session and cannot be changed without restarting R.

\code{\link[rJava]{.jinit}} is invoked (and Java locked) when you:
\itemize{
\item Explicitly call \code{library(rJava)}
\item Load any package that imports rJava (which auto-loads it as a dependency)
\item Even just use IDE autocomplete with \code{rJava::} (this triggers initialization!)
\item Call any rJava-dependent function
}

Once any of these happen, the Java version used by rJava for that session is locked in.
For command-line Java tools that don't use rJava, this function can be called at any
time to switch Java versions for subsequent system calls.
}

\keyword{internal}
