% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rJavaEnv-package.R
\docType{package}
\name{rJavaEnv-package}
\alias{rJavaEnv}
\alias{rJavaEnv-package}
\title{rJavaEnv: Java Environments for R Projects}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Install specific version of Java runtime environment at the R project level.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/e-kotov/rJavaEnv}
  \item \url{http://www.ekotov.pro/rJavaEnv/}
}

}
\author{
\strong{Maintainer}: Egor Kotov \email{kotov.egor@gmail.com} (\href{https://orcid.org/0000-0001-6690-5345}{ORCID})

}
\keyword{internal}
