% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate-deactivate.R
\name{rje_activate}
\alias{rje_activate}
\title{Activate a project-specific Java environment}
\usage{
rje_activate(project = NULL, restart = FALSE)
}
\arguments{
\item{project}{The project directory. Defaults to the current working directory.}

\item{restart}{If TRUE, restarts the R session after activation. Defaults to FALSE.}
}
\description{
Activate a project-specific Java environment
}
\examples{
\dontrun{
# Activate the current project
rJavaEnv::rje_activate()

# Activate a specific project
rJavaEnv::rje_activate(project = "~/projects/my-java-project")
}
}
