% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consent.R
\name{rje_consent_check}
\alias{rje_consent_check}
\title{Verify User Consent for rJavaEnv}
\usage{
rje_consent_check()
}
\value{
\code{TRUE} if consent is verified, otherwise an error is raised.
}
\description{
Ensure that the user has granted permission for rJavaEnv to manage files on their file system.
}
\details{
The function is based on the code of the \code{renv} package.
Copyright 2023 Posit Software, PBC
License: https://github.com/rstudio/renv/blob/main/LICENSE
}
\keyword{internal}
