% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate-deactivate.R
\name{rje_deactivate}
\alias{rje_deactivate}
\title{Deactivate a project-specific Java environment}
\usage{
rje_deactivate(project = NULL, clean = FALSE, restart = FALSE)
}
\arguments{
\item{project}{The project directory. Defaults to the current working directory.}

\item{clean}{If TRUE, also remove the rJavaEnv directory.}

\item{restart}{If TRUE, restarts the R session after deactivation. Defaults to FALSE.}
}
\description{
Deactivate a project-specific Java environment
}
\examples{
\dontrun{
# Deactivate the current project
rJavaEnv::rje_deactivate()

# Deactivate and clean the current project
rJavaEnv::rje_deactivate(clean = TRUE)
}
}
