% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consent.R
\name{rje_envvar_exists}
\alias{rje_envvar_exists}
\title{Helper for clean env var check}
\usage{
rje_envvar_exists(key)
}
\arguments{
\item{key}{The environment variable key to check.}
}
\description{
#' The function is based on the code of the \code{renv} package.
Copyright 2023 Posit Software, PBC
License: https://github.com/rstudio/renv/blob/main/LICENSE
}
\keyword{internal}
