% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_sdkman.R
\name{rje_sleep}
\alias{rje_sleep}
\title{Sleep wrapper (mockable in tests)}
\usage{
rje_sleep(seconds)
}
\arguments{
\item{seconds}{Number of seconds to sleep}
}
\description{
Sleep wrapper (mockable in tests)
}
\keyword{internal}
