% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_build_env.R
\name{set_java_build_env_vars}
\alias{set_java_build_env_vars}
\title{Helper function to set Java build environment variables}
\usage{
set_java_build_env_vars(java_home, quiet = FALSE)
}
\arguments{
\item{java_home}{The path to the desired \code{JAVA_HOME}.}

\item{quiet}{A \code{logical} value indicating whether to suppress messages. Can be \code{TRUE} or \code{FALSE}.}
}
\description{
Helper function to set Java build environment variables
}
\keyword{internal}
