% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_env.R
\name{set_java_env}
\alias{set_java_env}
\title{Set the JAVA_HOME and PATH environment variables to a given path}
\usage{
set_java_env(java_home)
}
\arguments{
\item{java_dir}{}
}
\value{
Nothing. Sets the JAVA_HOME and PATH environment variables.
}
\description{
Set the JAVA_HOME and PATH environment variables to a given path
}
