% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utilities.R
\name{urls_test_all}
\alias{urls_test_all}
\title{Test all Java URLs}
\usage{
urls_test_all()
}
\value{
A list with the results of testing all Java URLs.
}
\description{
Test all Java URLs
}
\keyword{internal}
