% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_java.R
\name{check_java_version_at_path}
\alias{check_java_version_at_path}
\title{Check installed Java ver at path}
\usage{
check_java_version_at_path(java_path)
}
\arguments{
\item{java_dir}{}
}
\value{
Java version, otherwise stops with an error.
}
\description{
Check installed Java ver at path
}
\examples{
#' \dontrun{
install_java("path/to/any-java-17-aarch64-macos-jdk.tar.gz")
}
}
