% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_env.R
\name{java_check_version_cmd}
\alias{java_check_version_cmd}
\title{Check installed Java version using terminal commands}
\usage{
java_check_version_cmd(java_home = NULL)
}
\arguments{
\item{java_home}{Path to Java home directory. If NULL, the function uses the JAVA_HOME environment variable.}
}
\value{
TRUE if successful, otherwise FALSE.
}
\description{
Check installed Java version using terminal commands
}
\examples{
\dontrun{
java_check_version_cmd()
}
}
