% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_env.R
\name{java_check_version_system}
\alias{java_check_version_system}
\title{Check and print Java path and version}
\usage{
java_check_version_system()
}
\value{
TRUE if successful, otherwise FALSE.
}
\description{
This function checks the Java executable path and retrieves the Java version,
then prints these details to the console.
}
\keyword{internal}
