% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_distributions_cache.R
\name{java_clear_distrib_cache}
\alias{java_clear_distrib_cache}
\title{Clear the Java distributions cache folder}
\usage{
java_clear_distrib_cache(
  cache_dir = tools::R_user_dir("rJavaEnv", which = "cache"),
  check = TRUE,
  confirm = TRUE
)
}
\arguments{
\item{cache_dir}{The cache directory to clear. Defaults to the user-specific data directory.}

\item{check}{Whether to list the contents of the cache directory before clearing it. Defaults to TRUE.}

\item{confirm}{Whether to ask for confirmation before clearing the cache. Defaults to TRUE.}
}
\value{
A message indicating whether the cache was cleared or not.
}
\description{
Clear the Java distributions cache folder
}
\examples{
\dontrun{
java_clear_distrib_cache()
}
}
