% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_download.R
\name{java_download}
\alias{java_download}
\title{Download a Java distribution}
\usage{
java_download(
  version = 21,
  distribution = "Corretto",
  dest_dir = tools::R_user_dir("rJavaEnv", which = "cache"),
  platform = platform_detect()$os,
  arch = platform_detect()$arch,
  verbose = TRUE
)
}
\arguments{
\item{version}{The Java version to download. If not specified, defaults to the latest LTS version.}

\item{distribution}{The Java distribution to download. If not specified, defaults to "Corretto".}

\item{dest_dir}{The destination directory to download the Java distribution to. Defaults to a user-specific data directory.}

\item{platform}{The platform for which to download the Java distribution. Defaults to the current platform.}

\item{arch}{The architecture for which to download the Java distribution. Defaults to the current architecture.}

\item{verbose}{Whether to print out information about the detected platform and architecture. Defaults to TRUE.}
}
\value{
The path to the downloaded Java distribution file.
}
\description{
Download a Java distribution
}
\examples{
\dontrun{
java_download(version = "17", distribution = "Corretto")
java_download(distribution = "Corretto")
java_download()
}
}
