% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_env.R
\name{java_env_set}
\alias{java_env_set}
\title{Set the JAVA_HOME and PATH environment variables to a given path}
\usage{
java_env_set(java_home, where = c("session", "project"), verbose = TRUE)
}
\arguments{
\item{java_home}{The path to the desired JAVA_HOME.}

\item{where}{Where to set the JAVA_HOME (and prepend to PATH): "session", "project", "global", or a combination of these options. Defaults to c("session", "project"). When set to "session", only the current R session is affected. When set to "project", the \code{.Rprofile} file in the project directory is updated and selected Java version will only be used in the current project. When set to "global", the global \code{.Rprofile} in user's home directory is updated, and the selected Java version will be used in all R sessions. If multiple values are provided, the function will set the JAVA_HOME and PATH in all specified locations.}

\item{verbose}{Whether to print detailed messages. Defaults to TRUE.}
}
\value{
Nothing. Sets the JAVA_HOME and PATH environment variables.
}
\description{
Set the JAVA_HOME and PATH environment variables to a given path
}
\examples{
\dontrun{
java_env_set("/path/to/java", c("session", "project"))
java_env_set("/path/to/java", "global")
}
}
