% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_env.R
\name{java_env_set_rprofile}
\alias{java_env_set_rprofile}
\title{Update the .Rprofile file in the project directory}
\usage{
java_env_set_rprofile(java_home, project_path = NULL)
}
\arguments{
\item{java_home}{The path to the desired JAVA_HOME.}

\item{project_path}{A \code{character} vector of length 1 containing the project directory where Java should be installed. If not specified or \code{NULL}, defaults to the current working directory.}
}
\description{
Update the .Rprofile file in the project directory
}
\keyword{internal}
