% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_env.R
\name{java_env_set_rprofile}
\alias{java_env_set_rprofile}
\title{Update the .Rprofile file in the project directory}
\usage{
java_env_set_rprofile(java_home)
}
\arguments{
\item{java_home}{The path to the desired JAVA_HOME.}
}
\description{
Update the .Rprofile file in the project directory
}
\keyword{internal}
