% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_env.R
\name{java_env_set_session}
\alias{java_env_set_session}
\title{Set the JAVA_HOME and PATH environment variables for the current session}
\usage{
java_env_set_session(java_home, quiet = FALSE, ._skip_rjava_check = FALSE)
}
\arguments{
\item{java_home}{The path to the desired JAVA_HOME.}

\item{quiet}{A \code{logical} value indicating whether to suppress messages. Can be \code{TRUE} or \code{FALSE}.}

\item{._skip_rjava_check}{Internal. If TRUE, skip the rJava initialization check.}
}
\description{
Set the JAVA_HOME and PATH environment variables for the current session
}
\keyword{internal}
