% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_env.R
\name{java_env_set_session}
\alias{java_env_set_session}
\title{Set the JAVA_HOME and PATH environment variables for the current session}
\usage{
java_env_set_session(java_home)
}
\arguments{
\item{java_home}{The path to the desired JAVA_HOME.}
}
\description{
Set the JAVA_HOME and PATH environment variables for the current session
}
\keyword{internal}
