% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_env.R
\name{java_env_unset}
\alias{java_env_unset}
\title{Unset the JAVA_HOME and PATH environment variables in the project .Rprofile}
\usage{
java_env_unset(project_path = NULL, quiet = FALSE)
}
\arguments{
\item{project_path}{A \code{character} vector of length 1 containing the project directory where Java should be installed. If not specified or \code{NULL}, defaults to the current working directory.}

\item{quiet}{A \code{logical} value indicating whether to suppress messages. Can be \code{TRUE} or \code{FALSE}.}
}
\value{
Nothing. Removes the JAVA_HOME and PATH environment variables settings from the project .Rprofile.
}
\description{
Unset the JAVA_HOME and PATH environment variables in the project .Rprofile
}
\examples{
# clear the JAVA_HOME and PATH environment variables in the specified project .Rprofile
java_env_unset(project_path = tempdir())
}
