% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_env.R
\name{java_env_unset}
\alias{java_env_unset}
\title{Unset the JAVA_HOME and PATH environment variables in the project .Rprofile}
\usage{
java_env_unset(project_path = NULL, verbose = TRUE)
}
\arguments{
\item{project_path}{The project directory where Java should be installed. If not specified or \code{NULL}, defaults to the current working directory.}

\item{verbose}{Whether to print detailed messages. Defaults to TRUE.}
}
\value{
Nothing. Removes the JAVA_HOME and PATH environment variables settings from the project .Rprofile.
}
\description{
Unset the JAVA_HOME and PATH environment variables in the project .Rprofile
}
\examples{
\dontrun{
# clear the JAVA_HOME and PATH environment variables in the specified project .Rprofile
java_env_unset(project_path = tempdir())
}
}
