% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_install.R
\name{java_install}
\alias{java_install}
\title{Install Java from a distribution file}
\usage{
java_install(
  java_path,
  project = NULL,
  autoset_java_path = TRUE,
  where = c("session", "project"),
  verbose = TRUE
)
}
\arguments{
\item{java_path}{The path to the Java distribution file.}

\item{project}{The project directory where Java should be installed. Defaults to the current working directory.}

\item{autoset_java_path}{Whether to set the JAVA_HOME and PATH environment variables to the installed Java directory. Defaults to TRUE.}

\item{where}{Where to set the JAVA_HOME: "session", "project", "global", or a combination. Defaults to c("session", "project").}

\item{verbose}{Whether to print detailed messages. Defaults to TRUE.}
}
\value{
The path to the installed Java directory.
}
\description{
Install Java from a distribution file
}
\examples{
\dontrun{
java_install("path/to/any-java-17-aarch64-macos-jdk.tar.gz")
}
}
