% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_manage.R
\name{java_list}
\alias{java_list}
\title{List the contents of the Java cache}
\usage{
java_list(
  type = c("project", "installed", "distrib"),
  output = c("data.frame", "vector"),
  verbose = FALSE,
  target_dir = NULL
)
}
\arguments{
\item{type}{The type of cache to list: "distrib", "installed", or "project".}

\item{output}{The format of the output: "data.frame" or "vector". Defaults to "data.frame".}

\item{verbose}{Whether to print detailed messages. Defaults to FALSE.}

\item{target_dir}{The cache directory to list. Defaults to the user-specific data directory for "distrib" and "installed", and the current working directory for "project".}
}
\value{
A data frame or character vector with the contents of the specified cache directory.
}
\description{
List the contents of the Java cache
}
\examples{
java_list("project")
java_list("installed")
java_list("distrib")

}
