% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_distributions_cache.R
\name{java_list_distrib_cache}
\alias{java_list_distrib_cache}
\title{List the contents of the Java distributions cache folder}
\usage{
java_list_distrib_cache(
  cache_dir = tools::R_user_dir("rJavaEnv", which = "cache"),
  verbose = FALSE
)
}
\arguments{
\item{cache_dir}{The cache directory to list. Defaults to the user-specific data directory.}

\item{verbose}{Whether to print detailed messages. Defaults to FALSE.}
}
\value{
A character vector with the contents of the cache directory.
}
\description{
List the contents of the Java distributions cache folder
}
\examples{
java_list_distrib_cache()

}
