% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_manage_distrib_cache.R
\name{java_list_distrib_cache}
\alias{java_list_distrib_cache}
\title{List the contents of the Java distributions cache folder}
\usage{
java_list_distrib_cache(
  cache_path = tools::R_user_dir("rJavaEnv", which = "cache"),
  output = c("data.frame", "vector"),
  verbose = FALSE
)
}
\arguments{
\item{cache_path}{The destination directory to download the Java distribution to. Defaults to a user-specific data directory.}

\item{output}{The format of the output: "data.frame" or "vector". Defaults to "data.frame".}

\item{verbose}{Whether to print detailed messages. Defaults to FALSE.}
}
\value{
A character vector with the contents of the cache directory.
}
\description{
List the contents of the Java distributions cache folder
}
\examples{
\dontrun{
java_list_distrib_cache()
}

}
