% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_manage_installed_cache.R
\name{java_list_installed_cache}
\alias{java_list_installed_cache}
\title{List the contents of the Java installations cache folder}
\usage{
java_list_installed_cache(
  output = c("data.frame", "vector"),
  verbose = FALSE,
  cache_path = tools::R_user_dir("rJavaEnv", which = "cache")
)
}
\arguments{
\item{output}{The format of the output: "data.frame" or "vector". Defaults to "data.frame".}

\item{verbose}{Whether to print detailed messages. Defaults to FALSE.}

\item{cache_path}{The cache directory to list. Defaults to the user-specific data directory. Not recommended to change.}
}
\value{
A data frame or character vector with the contents of the cache directory.
}
\description{
List the contents of the Java installations cache folder
}
\examples{
\dontrun{
java_list_installed_cache()
}

}
