% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_quick_install.R
\name{java_quick_install}
\alias{java_quick_install}
\title{Download and install and set Java in current working/project directory}
\usage{
java_quick_install(
  version = 21,
  distribution = "Corretto",
  platform = platform_detect()$os,
  arch = platform_detect()$arch,
  where = c("session", "project"),
  verbose = TRUE
)
}
\arguments{
\item{version}{The Java version to download. If not specified, defaults to the latest LTS version.}

\item{distribution}{The Java distribution to download. If not specified, defaults to "Corretto".}

\item{platform}{The platform for which to download the Java distribution. Defaults to the current platform.}

\item{arch}{The architecture for which to download the Java distribution. Defaults to the current architecture.}

\item{where}{Where to set the JAVA_HOME: "session", "project", "global", or a combination. Defaults to c("session", "project").}

\item{verbose}{Whether to print messages. Defaults to TRUE.}
}
\value{
Message indicating that Java was installed and set in the current working/project directory.
}
\description{
Download and install and set Java in current working/project directory
}
\examples{
\dontrun{
java_quick_install()
}
}
