% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utilities.R
\name{platform_detect}
\alias{platform_detect}
\title{Detect platform and architecture}
\usage{
platform_detect(quiet = TRUE)
}
\arguments{
\item{quiet}{A \code{logical} value indicating whether to suppress messages. Can be \code{TRUE} or \code{FALSE}.}
}
\value{
A list of length 2 with the detected platform and architecture.
}
\description{
Detect platform and architecture
}
\keyword{internal}
