% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rJavaEnv-package.R
\docType{package}
\name{rJavaEnv-package}
\alias{rJavaEnv}
\alias{rJavaEnv-package}
\title{rJavaEnv: Java Environments for R Projects}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Install specific version of Java runtime environment at the R project level. The goal of rJavaEnv is to manage multiple Java JDKs in R projects by automatingthe process of downloading, installing, and configuring Java environments on a per-project basis. This package is inspired by the renv \url{https://rstudio.github.io/renv/} package for managing R environments in R projects. You can request a specific Java Development Kit (JDK) in your project, and rJavaEnv will download and install the requested Java environment in a project-specific directory and set the PATH and JAVA_HOME for when you are using this project. Therefore, you can have different Java versions for different projects without contaminating your system with different Java versions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/e-kotov/rJavaEnv}
  \item \url{https://www.ekotov.pro/rJavaEnv/}
  \item Report bugs at \url{https://github.com/e-kotov/rJavaEnv/issues}
}

}
\author{
\strong{Maintainer}: Egor Kotov \email{kotov.egor@gmail.com} (\href{https://orcid.org/0000-0001-6690-5345}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Mauricio Vargas \email{mavargas11@uc.cl} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID}) [contributor]
}

}
\keyword{internal}
