% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_env.R
\name{set_java_build_env_vars}
\alias{set_java_build_env_vars}
\title{Set Java build environment variables for source installation}
\usage{
set_java_build_env_vars(java_home)
}
\arguments{
\item{java_home}{The path to the desired JAVA_HOME.}
}
\description{
Set Java build environment variables for source installation
}
\keyword{internal}
